refactor: :recycle: Refactored Files and added comments
Added comments and cleaned up code directorys for clearer understanding of how thing work
