--- packages/wp-sdk/src/components/logo/Logo.tsx
@@ -1,124 +1,0 @@
-import * as React from "react"
-import { SVGProps } from "react"
-const EverfundIcon = (props: SVGProps<SVGSVGElement>) => (
-  <svg
-    width={24}
-    height={24}
-    viewBox="0 0 24 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <g filter="url(#filter0_i_2010_2326)">
-      <rect width={24} height={24} rx={12} fill="#38C793" />
-      <rect width={24} height={24} rx={12} fill="white" fillOpacity={0.1} />
-      <g filter="url(#filter1_b_2010_2326)">
-        <mask id="path-2-inside-1_2010_2326" fill="white">
-          <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M19.2367 10.1056C19.2905 8.49638 17.9768 7.14826 16.3604 7.15289C15.3891 7.15564 14.8537 7.43745 14.4283 7.90774C14.1028 8.26752 13.5459 8.29669 13.1843 7.97288C12.8227 7.64907 12.7934 7.09491 13.1188 6.73512C13.913 5.85705 14.9428 5.40402 16.3553 5.40002M19.2367 10.1056C19.1854 11.6364 17.9208 12.8626 16.3693 12.8626C16.3723 12.8626 16.3719 12.8625 16.3683 12.8624C16.2976 12.8587 15.0006 12.7909 14.1982 11.5672C14.1943 11.5613 14.1904 11.5554 14.1863 11.5496C13.6584 10.7855 13.2942 10.2198 12.973 9.71891C12.9585 9.69622 12.944 9.67365 12.9296 9.65117C12.6262 9.17787 12.3502 8.74719 12.0075 8.26338L12.0074 8.26328C11.1789 7.09386 10.4654 6.35826 9.73606 5.93055C8.96913 5.48079 8.27164 5.42294 7.63058 5.42294C5.07263 5.42294 3 7.48607 3 10.0303C3 12.3619 4.45141 14.3344 6.04829 15.9191C6.86101 16.7257 7.7516 17.4704 8.58987 18.1459C8.85596 18.3604 9.11414 18.5658 9.36301 18.7638C9.91206 19.2006 10.4158 19.6015 10.8594 19.9835C11.5279 20.5631 12.5445 20.5391 13.1835 19.9032L15.7515 17.3481C16.0955 17.0058 16.0955 16.4509 15.7515 16.1086C15.4076 15.7663 14.8498 15.7663 14.5058 16.1086L11.9754 18.6263C11.5133 18.2302 10.9797 17.8056 10.4308 17.3689C10.1881 17.1757 9.94241 16.9802 9.69863 16.7838C8.8729 16.1184 8.03918 15.4192 7.29238 14.6781C5.77021 13.1674 4.76171 11.6296 4.76171 10.0303C4.76171 8.45395 6.04581 7.17582 7.63058 7.17582C8.13802 7.17582 8.46775 7.22146 8.84153 7.44065C9.25291 7.6819 9.79352 8.18071 10.5675 9.27311C10.8862 9.72314 11.1425 10.123 11.4505 10.6034C11.4628 10.6227 11.4753 10.6421 11.4878 10.6616C11.8111 11.1659 12.1867 11.7493 12.7288 12.5343C14.0814 14.587 16.2764 14.6155 16.3693 14.6155C18.8739 14.6155 20.9145 12.6366 20.9974 10.164C21.0845 7.56369 18.971 5.39254 16.3553 5.40002"
-          />
-        </mask>
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M19.2367 10.1056C19.2905 8.49638 17.9768 7.14826 16.3604 7.15289C15.3891 7.15564 14.8537 7.43745 14.4283 7.90774C14.1028 8.26752 13.5459 8.29669 13.1843 7.97288C12.8227 7.64907 12.7934 7.09491 13.1188 6.73512C13.913 5.85705 14.9428 5.40402 16.3553 5.40002M19.2367 10.1056C19.1854 11.6364 17.9208 12.8626 16.3693 12.8626C16.3723 12.8626 16.3719 12.8625 16.3683 12.8624C16.2976 12.8587 15.0006 12.7909 14.1982 11.5672C14.1943 11.5613 14.1904 11.5554 14.1863 11.5496C13.6584 10.7855 13.2942 10.2198 12.973 9.71891C12.9585 9.69622 12.944 9.67365 12.9296 9.65117C12.6262 9.17787 12.3502 8.74719 12.0075 8.26338L12.0074 8.26328C11.1789 7.09386 10.4654 6.35826 9.73606 5.93055C8.96913 5.48079 8.27164 5.42294 7.63058 5.42294C5.07263 5.42294 3 7.48607 3 10.0303C3 12.3619 4.45141 14.3344 6.04829 15.9191C6.86101 16.7257 7.7516 17.4704 8.58987 18.1459C8.85596 18.3604 9.11414 18.5658 9.36301 18.7638C9.91206 19.2006 10.4158 19.6015 10.8594 19.9835C11.5279 20.5631 12.5445 20.5391 13.1835 19.9032L15.7515 17.3481C16.0955 17.0058 16.0955 16.4509 15.7515 16.1086C15.4076 15.7663 14.8498 15.7663 14.5058 16.1086L11.9754 18.6263C11.5133 18.2302 10.9797 17.8056 10.4308 17.3689C10.1881 17.1757 9.94241 16.9802 9.69863 16.7838C8.8729 16.1184 8.03918 15.4192 7.29238 14.6781C5.77021 13.1674 4.76171 11.6296 4.76171 10.0303C4.76171 8.45395 6.04581 7.17582 7.63058 7.17582C8.13802 7.17582 8.46775 7.22146 8.84153 7.44065C9.25291 7.6819 9.79352 8.18071 10.5675 9.27311C10.8862 9.72314 11.1425 10.123 11.4505 10.6034C11.4628 10.6227 11.4753 10.6421 11.4878 10.6616C11.8111 11.1659 12.1867 11.7493 12.7288 12.5343C14.0814 14.587 16.2764 14.6155 16.3693 14.6155C18.8739 14.6155 20.9145 12.6366 20.9974 10.164C21.0845 7.56369 18.971 5.39254 16.3553 5.40002"
-          fill="url(#paint0_linear_2010_2326)"
-          fillOpacity={0.8}
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M19.2367 10.1056C19.2905 8.49638 17.9768 7.14826 16.3604 7.15289C15.3891 7.15564 14.8537 7.43745 14.4283 7.90774C14.1028 8.26752 13.5459 8.29669 13.1843 7.97288C12.8227 7.64907 12.7934 7.09491 13.1188 6.73512C13.913 5.85705 14.9428 5.40402 16.3553 5.40002M19.2367 10.1056C19.1854 11.6364 17.9208 12.8626 16.3693 12.8626C16.3723 12.8626 16.3719 12.8625 16.3683 12.8624C16.2976 12.8587 15.0006 12.7909 14.1982 11.5672C14.1943 11.5613 14.1904 11.5554 14.1863 11.5496C13.6584 10.7855 13.2942 10.2198 12.973 9.71891C12.9585 9.69622 12.944 9.67365 12.9296 9.65117C12.6262 9.17787 12.3502 8.74719 12.0075 8.26338L12.0074 8.26328C11.1789 7.09386 10.4654 6.35826 9.73606 5.93055C8.96913 5.48079 8.27164 5.42294 7.63058 5.42294C5.07263 5.42294 3 7.48607 3 10.0303C3 12.3619 4.45141 14.3344 6.04829 15.9191C6.86101 16.7257 7.7516 17.4704 8.58987 18.1459C8.85596 18.3604 9.11414 18.5658 9.36301 18.7638C9.91206 19.2006 10.4158 19.6015 10.8594 19.9835C11.5279 20.5631 12.5445 20.5391 13.1835 19.9032L15.7515 17.3481C16.0955 17.0058 16.0955 16.4509 15.7515 16.1086C15.4076 15.7663 14.8498 15.7663 14.5058 16.1086L11.9754 18.6263C11.5133 18.2302 10.9797 17.8056 10.4308 17.3689C10.1881 17.1757 9.94241 16.9802 9.69863 16.7838C8.8729 16.1184 8.03918 15.4192 7.29238 14.6781C5.77021 13.1674 4.76171 11.6296 4.76171 10.0303C4.76171 8.45395 6.04581 7.17582 7.63058 7.17582C8.13802 7.17582 8.46775 7.22146 8.84153 7.44065C9.25291 7.6819 9.79352 8.18071 10.5675 9.27311C10.8862 9.72314 11.1425 10.123 11.4505 10.6034C11.4628 10.6227 11.4753 10.6421 11.4878 10.6616C11.8111 11.1659 12.1867 11.7493 12.7288 12.5343C14.0814 14.587 16.2764 14.6155 16.3693 14.6155C18.8739 14.6155 20.9145 12.6366 20.9974 10.164C21.0845 7.56369 18.971 5.39254 16.3553 5.40002"
-          stroke="url(#paint1_linear_2010_2326)"
-          strokeWidth={10.5908}
-          mask="url(#path-2-inside-1_2010_2326)"
-        />
-      </g>
-    </g>
-    <defs>
-      <filter
-        id="filter0_i_2010_2326"
-        x={0}
-        y={-2.4}
-        width={24}
-        height={26.4}
-        filterUnits="userSpaceOnUse"
-        colorInterpolationFilters="sRGB"
-      >
-        <feFlood floodOpacity={0} result="BackgroundImageFix" />
-        <feBlend
-          mode="normal"
-          in="SourceGraphic"
-          in2="BackgroundImageFix"
-          result="shape"
-        />
-        <feColorMatrix
-          in="SourceAlpha"
-          type="matrix"
-          values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-          result="hardAlpha"
-        />
-        <feOffset dy={-2.4} />
-        <feGaussianBlur stdDeviation={2.4} />
-        <feComposite in2="hardAlpha" operator="arithmetic" k2={-1} k3={1} />
-        <feColorMatrix
-          type="matrix"
-          values="0 0 0 0 1 0 0 0 0 1 0 0 0 0 1 0 0 0 0.64 0"
-        />
-        <feBlend
-          mode="normal"
-          in2="shape"
-          result="effect1_innerShadow_2010_2326"
-        />
-      </filter>
-      <filter
-        id="filter1_b_2010_2326"
-        x={-18.6}
-        y={-16.2}
-        width={61.2}
-        height={58.2}
-        filterUnits="userSpaceOnUse"
-        colorInterpolationFilters="sRGB"
-      >
-        <feFlood floodOpacity={0} result="BackgroundImageFix" />
-        <feGaussianBlur in="BackgroundImageFix" stdDeviation={10.8} />
-        <feComposite
-          in2="SourceAlpha"
-          operator="in"
-          result="effect1_backgroundBlur_2010_2326"
-        />
-        <feBlend
-          mode="normal"
-          in="SourceGraphic"
-          in2="effect1_backgroundBlur_2010_2326"
-          result="shape"
-        />
-      </filter>
-      <linearGradient
-        id="paint0_linear_2010_2326"
-        x1={12}
-        y1={5.4}
-        x2={12}
-        y2={28.1453}
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop offset={0.313079} stopColor="white" />
-        <stop offset={1} stopColor="white" />
-      </linearGradient>
-      <linearGradient
-        id="paint1_linear_2010_2326"
-        x1={43.7717}
-        y1={-31.0068}
-        x2={-1.31494}
-        y2={12.8847}
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor="white" />
-        <stop offset={1} stopColor="white" />
-      </linearGradient>
-    </defs>
-  </svg>
-)
-export default EverfundIcon
