--- packages/wp-sdk/src/components/logo/index.tsx
@@ -1,3 +1,0 @@
-import Logo from './Logo';
-
-export default Logo;
