+++ packages/wp-sdk/src/donate-button/block.tsx
@@ -2,22 +2,20 @@
  * External dependencies
  */
 import clsx from 'clsx';
-import * as React from 'react';
 
 /**
  * Internal dependencies
  */
-import { BlockAttributes } from './types';
+import type { DonateButtonAttributes } from './types';
 
+type BlockProps = DonateButtonAttributes & { style?: React.CSSProperties };
 /**
  * This is the shared component for the donate button block, which renders both
  * in the editor and the frontend.
  * @param {BlockAttributes} props - The block props
  * @returns {JSX.Element} - The block component
  */
-export const Block = (
-	props: BlockAttributes & { style?: React.CSSProperties }
-): JSX.Element => {
+export const DonateButtonBlock = (props: BlockProps): JSX.Element => {
 	const { className, style, children, code } = props;
 
 	return (
@@ -35,5 +33,3 @@ export const Block = (
 		</button>
 	);
 };
-
-export default Block;
