+++ packages/wp-sdk/src/donation-form/types.ts
@@ -1,0 +1,6 @@
+export interface DonationFormAttributes {
+	code: string;
+	domain: string;
+	height?: string;
+	width?: string;
+}
